requirements.txt
opencv-python
mediapipe
numpy
scipy
matplotlib
tk
Pillow
